Share cgroup namespace in pod
`podman help pod create` displays `cgroup` as one of default values of
`--share` flag, which specifies shared namespaces in a pod. However,
`cgroup` is not listed in `SharedNameSpaces` of a pod with default
values. Even if `cgroup` is explicitly specified in `--share` flag,
`cgroup` is not listed in pod inspection and actually containers in a
pod don't share a cgroup namespace.

There are two similar parameters for cgroups in a pod. When `cgroup`
is specified in `--share`, `PodConfig.UsePodCgroup` is set to `true`.
For `SharedNameSpaces` in pod inspection, `PodConfig.UsePodCgroupNS` is
referred to.

In order to share `cgroup` namespace expectedly,
`PodConfig.UsePodCgroupNS` should be set to `true` via
`WithPodCgroup()`. Regarding `PodConfig.UsePodCgroup`, this should be
always `true` via `WithPodCgroups()` as it was in Podman 3.3.

Signed-off-by: Hironori Shiina <shiina.hironori@jp.fujitsu.com>