make reference --filter Docker compatible
By vendoring in c/common@main.  Notice that this a breaking change
compared to how Podman filtered images before.  Now the behavior is
identical to the one of Docker.

Most notably, to match an image, the domain and repository must be
specified as well.  Previously, `reference=foo` matched *any* image
named `foo`.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>