[WIP] VirtualBox provider
Add VirtualBox as a provider for `podman machine`

Add optional argument `--type` to specify provider
type for all `podman machine` subcommands (start,
stop, ls, ssh, rm, set). Keep old behaviour when
`--type` is not provided.

At now, VirtualBox provider has limited support for
some reasons such as:
- Fedora CoreOS does not provide images for VirtualBox,
  and it needs to use user-defined/custom image
- It was checked only on Fedora 35 and macOS 12.1 (Intel),
  no any sort of warranty for others

**Motivation for this PR**:  Allow using VBox when other
options (like QEMU) are not available. For instance,
sometimes there are cases when QEMU couldn't be used
or not allowed for usage, it could be for some a company
restriction and/or absence of security team approval.

Progress (what was done, what will be):
- [x] MVP version of VirtualBox provider
- [x] podman.socket
- [ ] port forwarding
- [ ] volume mounting
- [ ] Updated `podman machine` cli
- [ ] Documented

[NO NEW TESTS NEEDED]

Signed-off-by: esendjer <esendjer@gmail.com>