troubleshooting.md: tweak subuid paragraph, encryption
Expand the bit about needing to allocate UIDs so that we don't appear to
imply that adding a range of 10000 IDs to /etc/subuid will allow people
to use UID 1000000, which isn't in the range that we'd map a range of
that size to.

TLS is an acronym, so capitalize when we're talking about the protocol.

TLS verification is about encryption, not authentication.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>